refactor: move search diagnostics observer from internal directory to separate ftsstats package
