+++ pkg/ftsstats/search_stats.go
@@ -1,4 +1,4 @@
-package observability
+package ftsstats
 
 import (
 	"crypto/sha256"
