+++ pkg/ftsstats/search_stats_test.go
@@ -1,4 +1,4 @@
-package observability
+package ftsstats
 
 import (
 	"errors"
